Paul Davis <pbd@op.net>
This set of patches:

* fixes a problem with initialization when firmware was
already loaded (we ended up with random channel counts)
* adds the snd_line_outs_monitor option
- if enabled, all input and playback streams are
routed with unity gain to the line outs
* adds the snd_force_firmware option
- if enabled, I/O box firmware is reloaded whether
its already loaded or not

There are also a couple of minor corrections.

The driver/hardware still suffers from a problem that requires 1 warm
reboot (or PCMCIA card eject/reinsert) after a cold boot to get output
working. I am continuing to work on a solution to the problem.