- SDPIF input support, only 48khz - PCM multichannel bug fix. - Minor cleanups and bugfixes.
mixer changes:
- SPDIF Output Switch - mutes, unmutes SPDIF output
- SPDIF Input Switch - mutes, unmutes SPDIF input
- SPDIF Input Volume - adjust the volume on SPDIF input
- DAC capture Switch - capture what goes out to the CODEC, the primary
use of this is to capture the input from the SPDIF,
but it will capture all PCM playback as well.
- ADC capture Switch - capture what goes analog input to the CODEC like,
mic line in, aux in. etc ...
- Optical/Coaxial SPDIF Input Switch - only Hercules Game Theater XP,
switch between SPDIF toslink and coxial source, have not tested
the coaxial input yet. (dont have any HW with SPDIF coaxial output)