mega-patch by Thomas Charbonnel <thomas@undata.org>:
The driver now depends on the hwdep interface, the snd-hdsp entry of the
alsa-kernel/core/Makefile should be edited to add this dependency.

This driver should solve issues with the rev11 firmwares and the
hdsp9652 cards.

The card won't work until the firmware is uploaded with the hdspload
tool. In case the firmware gets lost (e.g. power supply failure), just
try to access the card again and the firmware will get uploaded back
from kernel cached version, or run the loader again. Note that at this
point the card will be reset to its default values (your mixer settings
will be lost).