Paul Davis <paul@linuxaudiosystems.com>
this patch fixes some basic problems with the hdsp driver with respect
to the hdsp9652 card. it also cleans up a few minor issues with naming
in the driver, and slightly rationalizes initialization to involve
the minimum of special-casing for the hdsp9652.

the basic problem with the hdsp9652 was related to 8 bit versus 32 bit
offsets when addressing the mixer memory. once this was fixed,
everything worked. this driver continues to work fine on my
pci+digiface unit as well.