Steve deRosier <derosier@pianodisc.com>:
* There is a user selectable flag "droponfull".  Set to 1 and
any new bytes delivered to the driver after the buffer fills
up will be discarded until the buffer is able to flush some
bytes.
* If droponfull==0 (or is not set, the default is 0), the driver
proceeds to block further input by not calling
snd_rawmidi_transmit_ack() and aborting the attempt.  It will
try again later.
* I've redone a bit of the interface for the buffer routines.
This was done to support the proper blocking/non-blocking methods
as spelled out above, and to try to protect the buffer data a bit.