Clean up the buffer management in the PCM runtime record.
The buffer-allocation record is hold in runtime instance, so that
it can be checked more cleanly.

dma_private is removed from runtime (it's used for SG-buffers).
The macro snd_pcm_substream_sgbuf() should be used instead of direct
access to the pointer, to retrieve the sgbuf struct from the PCM
substream instance.