Signed-off-by: Chris Cannam <cannam@all-day-breakfast.com>
snd_pcm_timer_resolution_change():

Right, that function doesn't work well for 44100/1024 in 32 bits -- it
ends up trying to calculate 1e7 * 1024 / 441 and having to divide
both numerator and denominator by 4 (thus losing the rather crucial 1
at the end of 441) before it can do the calculation without overflow.

Attached is a patch against 1.0.5a that gets better results in this
instance, by leaving the denominator alone and instead doubling the
result back up by the same number of times as the multiplier had to
be halved by.