ALSA Driver for Sparc DBRI chips
Initial DBRI driver support by Martin Habets <errandir_news@mph.eclipse.co.uk>:

Hi,

I have been working on an ALSA driver for the DBRI/CS4215 chip
combo found in Sparc machines. It is a port of the 2.5 OSS driver.

Things are starting to work, but since this is my first major
undertaking I'm looking for your comments. Ask early, ask often...
At this point I expect playback to work to all possible outputs (there
is some clipping on 16 bit formats). I have not been able to get recording
output in a file, though it does seem to catch data.

So I look for anything from coding style to design issues to spelling to
'it worked' to 'I get an error' to the-meaning-of-the-universe.
Just keep the phasers on stun, please :)