Hotplug firmware loader support
The hogplut fw loader is supported by vxpocket, vxp440, vx2222 and mixart
drivers.  The old ALSA fw loader is still supported for built-in kernels.
The hotplug fw is disabled on 2.2/2.4 kernels (although the recent 2.4
could support it).
To use the hotplug, the new firmware data must be installed beforehand
from the latest alsa-tools package.

The experimental suspend/resume for vxpocket, vxp440 and vx222 are added,
too.