Interpret negative index as bitmask of permissible indexes
Currently arguments of the "index" option from 0 through SNDRV_CARDS-1
force a module to take the specified index. Index -1 makes the module
take the first available index. This patch extends this convention so
that a negative index value is interpreted as a bitmask of the
permitted indexes. Special cases:

-1 0xffffffff 0 and up
-2 0xfffffffe 1 and up
-4 0xfffffffc 2 and up
...

The patch includes also corrections of ALSA-Configuration.txt document.

Signed-off-by: Thomas Hood <jdthood@yahoo.co.uk>