CMI8768 patch
Hi,

I made a patch for CM8768, which has the same PCI ID but there are
several HW diffecece, aas listed:

1. 8768 has no PCM volume control.
2. The ADC of 8768 can only record in 44.1kHz or 48kHz.
3. 8768 can support up to 8 channels.

I made change for item 2 and 3, I want to use the softvol plugin but
don't know how to. The driver just don't generate the PCM volume slider
for now.

Signed-off-by: ChenLi Tien <cltien@cmedia.com.tw>