replace spdif frequency control with iec958 control
The appended patch replaces the manual spdif frequency control in the
mixer with code that use the standard iec958 interface. The patch is
nearly the same as the one Raymond submitted, except it gets rid of the
identical *_default_* versions of the functions, and has a tiny bit of
cleanup.
I have only tested it a little bit because I don't currently
have the hardware to actually test it properly. However, it looks
harmless to me.

Patch-by: Raymond
Signed-off-by: Jeff Muizelaar <muizelaar@rogers.com>