RME HDSP MADI pci naming bug
I discovered a trivial bug where the snd-hdsp and snd-hdspm do not
play together.  They cannot be loaded as modules simultaneously and create
detection problems if compiled as static modules in a kernel.  The
problem come from the sysfs while registering their PCI names, which
happens to be the same (oops).  Therefore the first module to get loaded
stayed in, but the secound one would be blocked...not anymore ;)

From: Joachim Deguara <dga@idmt.fraunhofer.de>