hda-codec - Add support of more models with ALC codecs
Merged the work of pshou <pshou@realtek.com.tw> for the support of
more models with ALC codecs: ALC880 ASUS, Uniwill, FSC1734, generic 6-stack,
and ALC260 HP.  Tests with the real hardwares are appreciated.

The codec patch is cleaned up:  The preset configuration of codecs are
stored in the table and copied to the spec instance.

Added/fixed comments.