cleanup and typo-correction
This patch changes, adds and remove some comments, which will
make now more sense and fit on a 80-char line. It also changes
the order of snd_power_wait() to make the file more readable.
It removes the device.c comment in front of _snd_minor,
cause snd_minor has nothing to do with device.c.
The both typos in the kernel-docs were corrected too.

Signed-off-by: Henrik Kretzschmar <henne@nachtwindheim.de>