snd-ad1889: simplify DMA buffer handling
Instead of manually programming DMA for each period, just put the
buffer size into the count registers (BC, CC) and the period size into
the interrupt count registers (IC, IB).  This removes the need to
reprogram any of the chip's registers after the stream has been
started, and allows buffer sizes that are not a multiple of the period
size.