snd_powermac: Add ID for Spring 2005 17" Powerbook
The audio chip in my Spring 2005 17" PowerBook was incorrectly
recognized as an AWACS chip.  This adds the chip ID to the
snd_powermac driver such that it is recognized as a Toonie (I don't
know if that's correct, but it's the only one that makes it work at
all). and sorts the ID lists numerically.  NOTE:  This chip is only
minimally supported at this point; it has system beep support and
very low volume speaker output, and that's about it.

Signed-off-by: Kyle Moffett <mrmacman_g4@mac.com>