Echoaudio - new firmware
This patchset adds support for the new firmware for "3G" cards. It fixes
clicks/pops during simultaneous audio and midi i/o and it uses the same DSP
firmware for both the Gina3G and the Layla 3G.

I announced this patch two months ago asking for someone to test it on a
Layla3G, but nobody cared. I tested it a lot on the Gina3G. The patch is
against the latest anonymous CVS version.

Signed-off-by: Giuliano Pochini <pochini@shiny.it>