intel8x0 - Fix duplicate ac97_quirks entry From: Ulrich Mueller <ulm@kph.uni-mainz.de>
in recent -mm kernels additional quirks for ac97 hardware in HP laptops
have been added.  However, now the list in intel8x0.c contains a duplicate,
since the HP nx6110 and nc6120 have identical subdevice ids.

This was introduced in -mm1 by the following patch:
add-new-quirk-for-devices-with-mute-leds-and-separate-headphone-volume.patch

Since the HP nx6110 and nc6120 are almost identical, both entries
should really be combined, as in the following patch. I have checked
that AC97_TUNE_HP_MUTE_LED is the right thing to do.

Signed-off-by: Andrew Morton <akpm@osdl.org>