#2/4 for Zoom Video - resolve number of record sources
Second of 4 es18xx.c patches culminating in Zoom Video support.

This patch changes the 'record source' mux routines to reflect the fact
that not all of the
supported chipsets have 8 possible inputs. Some have 4 and some have 5.

Testing:
This work was initially done on the source from the Debian Sarge ALSA
package, then tested
on an ES1879 and an ES1878 machine. Patches were created against the
Sarge code and then edited
to apply correctly to the ALSA cvs code. Lastly the patched ALSA cvs
code was test for
successful compilation. No additional testing was done on the ALSA cvs
version.

Signed-off-by: Mark Salazar <markTheCoder@justmyself.net>