HDA/ALC260: 1/7 - Fix test model input mux label From: Jonathan Woithe <jwoithe@physics.adelaide.edu.au>
This patch fixes input mux labels used for the ALC260 "test" model.  This is
needed to avoid confusion which comes about because the two ADCs in the
ALC260 have slightly different mappings for their input selectors. Since
this is just the test model it's sufficient to simply report both options
where they exist.  If a model comes along for which this becomes an issue,
the ALC260 input mux code will have to be extended to allow different mux
layouts for different ADCs.