HDA/ALC260: 5/7 - add "acer" model From: Jonathan Woithe <jwoithe@physics.adelaide.edu.au>
This patch adds a new "acer" model.  It is based on work by Rimas and many
other dedicated Acer owners over the past few weeks (see bug number
1618) - kudos to them. This adds support for a variety of Acer laptops,
although we are still in the process of collecting pci/subsystem IDs to add.
There are still some potentially outstanding issues: there are reports that
the CD control might not yet be functional for example.  However, the time
is probably right to get a "first cut" into the kernel which can be refined
as more test reports come in.

From the reports by various Acer owners, this code allows them to record
from internal mics and external jacks.  Playback also seems to work to all
external jacks and the internal speaker.  The "beep" control doesn't appear
functional at the moment.

This patch depends on the changes made in patch 3/7.