HDA/ALC260: 7/7 - add SPDIF enable to test model From: Jonathan Woithe <jwoithe@physics.adelaide.edu.au>
This patch adds mixer controls to the "test" ALC260 model which allow the
user to selectively enable or disable the SPDIF output pins.  This should
assist people identify digital outputs on machines which bring them to the
outside world.

Note that while the patch *should* work, I cannot personally verify it since
my laptop doesn't bring the SPDIF lines out.

As for the GPIO switches added in patch 4, these controls are currently
only compiled in if debug mode is selected.