ice1712 - Fix wordclock status on Delta1010LT From: Doug McLain <doug@nostar.net>
I posted this patch to bug 1806 a while back, and have been awaiting a
reply or commit.  It currently reports the Locked/No Signal text
backwards in envy24control, since i was using an older version of
envy24control when I wrote the original code.  The Locked/No Signal test
was recently reversed in envy24control cvs, so the test in my code needs
to be reversed as well.  Here is the patch, once again.